chore: split near oom file
